+++ fuzz/fuzz_targets/fuzz_fuel.rs
@@ -2,8 +2,8 @@
 
 use libfuzzer_sys::fuzz_target;
 
-use zfuel::fuel::ZFuel;
+use zfuel::fuel::{Precision, ZFuel};
 
 fuzz_target!(|data: i64| {
-    ZFuel::new(data);
+    ZFuel::new(data, Precision::DEFAULT);
 });