+++ src/lib.rs
@@ -4,3 +4,4 @@ pub mod error;
 pub mod fee;
 pub mod fraction;
 pub mod fuel;
+mod precision;